requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
geographiclib==1.52
geopy==2.2.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
Shapely==1.8.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
